requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
openai
faiss-cpu
python-dotenv
tiktoken
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
docx2txt
pypdf
sqlmodel
pymupdf
python-docx 
requests
python-multipart